+++ xclim/testing/tests/test_indices.py
@@ -216,7 +216,8 @@ class TestAgroclimaticIndices:
     @pytest.mark.parametrize(
         "method, end_date, deg_days, max_deg_days",
         [
-            ("gladstones", "11-01", 1127.78, 1926.0),
+            ("gladstones", "11-01", 1102.1, 1926.0),
+            ("jones", "11-01", 1203.4, 2179.1),
             ("icclim", "10-01", 915.0, 1647.0),
         ],
     )
@@ -271,17 +272,28 @@ class TestAgroclimaticIndices:
             freq="YS",
         )
 
-        np.testing.assert_allclose(
-            bedd, np.array([deg_days, deg_days, deg_days, np.NaN])
-        )
-        np.testing.assert_array_equal(
-            bedd_hot, [max_deg_days, max_deg_days, max_deg_days, np.NaN]
-        )
+        if method == "jones":
+            np.testing.assert_array_less(
+                bedd[1], bedd[0]
+            )  # Leap-year has slightly higher values
+            np.testing.assert_allclose(
+                bedd, np.array([deg_days, deg_days, deg_days, np.NaN]), rtol=3e-4
+            )
+            np.testing.assert_allclose(
+                bedd_hot, [max_deg_days, max_deg_days, max_deg_days, np.NaN], rtol=0.15
+            )
 
-        if method == "gladstones":
-            np.testing.assert_array_less(bedd, bedd_high_lat)
-        if method == "icclim":
-            np.testing.assert_array_equal(bedd, bedd_high_lat)
+        else:
+            np.testing.assert_allclose(
+                bedd, np.array([deg_days, deg_days, deg_days, np.NaN])
+            )
+            np.testing.assert_array_equal(
+                bedd_hot, [max_deg_days, max_deg_days, max_deg_days, np.NaN]
+            )
+            if method == "gladstones":
+                np.testing.assert_array_less(bedd, bedd_high_lat)
+            if method == "icclim":
+                np.testing.assert_array_equal(bedd, bedd_high_lat)
 
     def test_cool_night_index(self):
         ds = open_dataset("cmip5/tas_Amon_CanESM2_rcp85_r1i1p1_200701-200712.nc")
@@ -320,6 +332,37 @@ class TestAgroclimaticIndices:
         )
         lti = lti.groupby_bins(lti.lon, 1).mean().groupby_bins(lti.lat, 5).mean()
         np.testing.assert_array_almost_equal(lti[0].transpose(), np.array([values]), 2)
+
+    @pytest.mark.parametrize(
+        "method, end_date, values",
+        [
+            ("smoothed", "10-01", 1702.87),
+            ("icclim", "11-01", 1983.53),
+            ("jones", "10-01", 1729.12),
+            ("jones", "11-01", 2219.51),
+        ],
+    )
+    def test_huglin_index(self, method, end_date, values):
+        ds = open_dataset("cmip5/tas_Amon_CanESM2_rcp85_r1i1p1_200701-200712.nc")
+        ds = ds.drop_isel(time=0)  # drop time=2006/12 for one year of data
+
+        tasmax, tasmin = ds.tas + 15, ds.tas - 5
+        # It would be much better if the index would interpolate to daily from monthly data intelligently.
+        tasmax, tasmin = (
+            tasmax.resample(time="1D").interpolate("cubic"),
+            tasmin.resample(time="1D").interpolate("cubic"),
+        )
+        tasmax.attrs["units"], tasmin.attrs["units"] = "K", "K"
+
+        hi = xci.huglin_index(
+            tasmax=tasmax,
+            tasmin=tasmin,
+            lat=ds.lat,
+            method=method,
+            end_date=end_date,
+        )
+
+        np.testing.assert_almost_equal(np.mean(hi), values, 2)
 
 
 class TestDailyFreezeThawCycles:
@@ -1941,23 +1984,32 @@ def test_humidex(tas_series):
     tas = tas_series([15, 25, 35, 40])
     tas.attrs["units"] = "C"
 
-    dtas = tas_series([10, 15, 25, 25])
-    dtas.attrs["units"] = "C"
+    dtps = tas_series([10, 15, 25, 25])
+    dtps.attrs["units"] = "C"
 
     # expected values from https://en.wikipedia.org/wiki/Humidex
     expected = np.array([16, 29, 47, 52]) * units.degC
 
     # Celcius
-    hc = xci.humidex(tas, dtas)
+    hc = xci.humidex(tas, dtps)
     np.testing.assert_array_almost_equal(hc, expected, 0)
 
     # Kelvin
-    hk = xci.humidex(convert_units_to(tas, "K"), dtas)
+    hk = xci.humidex(convert_units_to(tas, "K"), dtps)
     np.testing.assert_array_almost_equal(hk, expected.to("K"), 0)
 
     # Fahrenheit
-    hf = xci.humidex(convert_units_to(tas, "fahrenheit"), dtas)
+    hf = xci.humidex(convert_units_to(tas, "fahrenheit"), dtps)
     np.testing.assert_array_almost_equal(hf, expected.to("fahrenheit"), 0)
+
+    # With relative humidity
+    hurs = xci.relative_humidity(tas, dtps, method="bohren98")
+    hr = xci.humidex(tas, hurs=hurs)
+    np.testing.assert_array_almost_equal(hr, expected, 0)
+
+    # With relative humidity and Kelvin
+    hk = xci.humidex(convert_units_to(tas, "K"), hurs=hurs)
+    np.testing.assert_array_almost_equal(hk, expected.to("K"), 0)
 
 
 @pytest.mark.parametrize(
@@ -1995,6 +2047,108 @@ def test_wind_chill(tas_series, sfcWind_series):
 
     out = xci.wind_chill_index(tas=tas, sfcWind=sfcWind, method="US")
     assert out[-1].isnull()
+
+
+class TestClausiusClapeyronScaledPrecip:
+    def test_simple(self):
+        pr_baseline = xr.DataArray(
+            np.arange(4).reshape(1, 2, 2),
+            dims=["time", "lat", "lon"],
+            coords={"time": [1], "lat": [-45, 45], "lon": [30, 60]},
+            attrs={"units": "mmday"},
+        )
+        tas_baseline = xr.DataArray(
+            np.arange(4).reshape(1, 2, 2),
+            dims=["time", "lat", "lon"],
+            coords={"time": [1], "lat": [-45, 45], "lon": [30, 60]},
+            attrs={"units": "C"},
+        )
+        tas_future = xr.DataArray(
+            np.arange(40).reshape(10, 2, 2),
+            dims=["time_fut", "lat", "lon"],
+            coords={"time_fut": np.arange(10), "lat": [-45, 45], "lon": [30, 60]},
+            attrs={"units": "C"},
+        )
+        delta_tas = tas_future - tas_baseline
+        delta_tas.attrs["units"] = "delta_degC"
+        out = xci.clausius_clapeyron_scaled_precipitation(delta_tas, pr_baseline)
+
+        np.testing.assert_allclose(
+            out.isel(time=0),
+            [
+                [
+                    [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
+                    [
+                        1.0,
+                        1.31079601,
+                        1.71818618,
+                        2.25219159,
+                        2.95216375,
+                        3.86968446,
+                        5.07236695,
+                        6.64883836,
+                        8.7152708,
+                        11.42394219,
+                    ],
+                ],
+                [
+                    [
+                        2.0,
+                        2.62159202,
+                        3.43637236,
+                        4.50438318,
+                        5.9043275,
+                        7.73936892,
+                        10.14473391,
+                        13.29767673,
+                        17.4305416,
+                        22.84788438,
+                    ],
+                    [
+                        3.0,
+                        3.93238803,
+                        5.15455854,
+                        6.75657477,
+                        8.85649125,
+                        11.60905339,
+                        15.21710086,
+                        19.94651509,
+                        26.1458124,
+                        34.27182657,
+                    ],
+                ],
+            ],
+        )
+
+    def test_workflow(self, tas_series, pr_series):
+        """Test typical workflow."""
+        n = int(365.25 * 10)
+        tref = tas_series(np.random.rand(n), start="1961-01-01")
+        tfut = tas_series(np.random.rand(n) + 2, start="2051-01-01")
+        pr = pr_series(np.random.rand(n) * 10, start="1961-01-01")
+
+        # Compute climatologies
+        with xr.set_options(keep_attrs=True):
+            tref_m = tref.mean(dim="time")
+            tfut_m = tfut.mean(dim="time")
+            pr_m = pr.mean(dim="time")
+
+        delta_tas = tfut_m - tref_m
+        delta_tas.attrs["units"] = "delta_degC"
+        pr_m_cc = xci.clausius_clapeyron_scaled_precipitation(delta_tas, pr_m)
+        np.testing.assert_array_almost_equal(pr_m_cc, pr_m * 1.07 ** 2, 1)
+
+        # Compute monthly climatologies
+        with xr.set_options(keep_attrs=True):
+            tref_mm = tref.groupby("time.month").mean()
+            tfut_mm = tfut.groupby("time.month").mean()
+            pr_mm = pr.groupby("time.month").mean()
+
+        delta_tas_m = tfut_mm - tref_mm
+        delta_tas_m.attrs["units"] = "delta_degC"
+
+        pr_mm_cc = xci.clausius_clapeyron_scaled_precipitation(delta_tas_m, pr_mm)
+        np.testing.assert_array_almost_equal(pr_mm_cc, pr_mm * 1.07 ** 2, 1)
 
 
 class TestPotentialEvapotranspiration: